Added messageText option for error reporting.
I had an issue with typescript compilation (awesome-typescript-loader for webpack) and all that was reported in the log was this:

    ERROR in ./src/app/index.module.ts
    Module build failed: [object Object]

when I debugged it, I noticed that it uses `messageText` instead of `message` for error message, so this is how I found out the real cause of the problem. The problem was not a typical TS compilation error, but due to this error message it took me a while to figure out the cause. Perhaps it's more the awesome-typescript-loader issue, but it's not up to me to decide.
Thanks.
Lukas